mlops74/requirements_fast.txt
click==8.1.7
hydra-core==1.3.2
Markdown==3.5.2
numpy==1.26.3
omegaconf==2.3.0
opencv_python==4.9.0.80
Pillow==10.2.0
pytest==7.4.4
scikit_learn==1.4.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
wandb==0.16.2
uvicorn>=0.15.0,<0.16.0